requirements.txt
requests==2.31.0
firebase-admin==6.5.0
google-generativeai==0.8.2
python-dotenv==1.0.1
pillow==10.4.0
pypdf2==3.0.1
python-docx==1.1.2
chardet==5.2.0
timezonefinder==6.5.9
numpy==1.26.4
flask==3.0.3
werkzeug==3.0.4